idea clean

+++ src/main/java/com/science/gtnl/common/machine/hatch/BeamlinePipeMirror.java
@@ -63,7 +63,7 @@ public class BeamlinePipeMirror extends MTEBeamlinePipe {
                         }
 
                         if (canConnect) {
-                            mConnections |= 1 << dir.ordinal();
+                            mConnections |= (byte) (1 << dir.ordinal());
                             connectedSides[connectionCount] = dir;
                             connectionCount++;
                         }
